+++ src/main/java/com/springboot/bookMyShow/services/MovieService.java
@@ -7,7 +7,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.springboot.bookMyShow.Entity.Admin;
 import com.springboot.bookMyShow.Entity.Movie;
+import com.springboot.bookMyShow.dao.AdminDao;
 import com.springboot.bookMyShow.dao.MovieDao;
 import com.springboot.bookMyShow.util.ResponceStructure;
 
@@ -17,15 +19,23 @@ public class MovieService {
 	@Autowired
 	MovieDao mDao;
 	
-	public ResponseEntity<ResponceStructure<Movie>> saveMovie(Movie movie)
+	@Autowired
+	AdminDao aDao;
+	
+	public ResponseEntity<ResponceStructure<Movie>> saveMovie(Movie movie ,String aEmail,String aPassword)
 	{
-		ResponceStructure<Movie> str=new ResponceStructure<Movie>();
-		
-		str.setMessage(movie.getMTitle()+" Movie has saved");
-		str.setStatus(HttpStatus.CREATED.value());
-		str.setData(mDao.saveMovie(movie));
-		
-		return new ResponseEntity<ResponceStructure<Movie>>(str,HttpStatus.CREATED);
+		Admin exa=aDao.adminLogin(aEmail, aPassword);
+		if(exa!=null)
+		{
+			ResponceStructure<Movie> str=new ResponceStructure<Movie>();
+			
+			str.setMessage(movie.getMTitle()+" Movie has saved");
+			str.setStatus(HttpStatus.CREATED.value());
+			str.setData(mDao.saveMovie(movie));
+			
+			return new ResponseEntity<ResponceStructure<Movie>>(str,HttpStatus.CREATED);
+		}
+		return null;
 		
 	}
 	
@@ -44,36 +54,47 @@ public class MovieService {
 		return null;
 	}
 	
-	public ResponseEntity<ResponceStructure<Movie>> deleteMovie(int mId)
+	public ResponseEntity<ResponceStructure<Movie>> deleteMovie(int mId ,String aEmail,String aPassword)
 	{
-		ResponceStructure<Movie> str= new ResponceStructure<Movie>();
-		
-		Movie m= mDao.findMovie(mId);
-		
-		if(m!=null)
+		Admin exa=aDao.adminLogin(aEmail, aPassword);
+		if(exa!=null)
 		{
-			str.setMessage(m.getMTitle()+"Movie has Deleted");
-			str.setStatus(HttpStatus.OK.value());
-			str.setData(mDao.deleteMovie(mId));
+			ResponceStructure<Movie> str= new ResponceStructure<Movie>();
 			
-			return new ResponseEntity<ResponceStructure<Movie>>(str,HttpStatus.OK);
+			Movie m= mDao.findMovie(mId);
+			
+			if(m!=null)
+			{
+				str.setMessage(m.getMTitle()+"Movie has Deleted");
+				str.setStatus(HttpStatus.OK.value());
+				str.setData(mDao.deleteMovie(mId));
+				
+				return new ResponseEntity<ResponceStructure<Movie>>(str,HttpStatus.OK);
+			}
+			return null;
 		}
 		return null;
 		
+		
 	}
 	
-	public ResponseEntity<ResponceStructure<Movie>> updateMovie(Movie movie,int mId)
+	public ResponseEntity<ResponceStructure<Movie>> updateMovie(Movie movie,int mId ,String aEmail,String aPassword)
 	{
-		ResponceStructure<Movie> str= new ResponceStructure<Movie>();
-		
-		Movie m=mDao.findMovie(mId);
-		if(m!=null)
+		Admin exa=aDao.adminLogin(aEmail, aPassword);
+		if(exa!=null)
 		{
-			str.setMessage(movie.getMTitle()+" movie has updated");
-			str.setStatus(HttpStatus.OK.value());
-			str.setData(mDao.updateMovie(movie, mId));
+			ResponceStructure<Movie> str= new ResponceStructure<Movie>();
 			
-			return new ResponseEntity<ResponceStructure<Movie>>(str, HttpStatus.OK);
+			Movie m=mDao.findMovie(mId);
+			if(m!=null)
+			{
+				str.setMessage(movie.getMTitle()+" movie has updated");
+				str.setStatus(HttpStatus.OK.value());
+				str.setData(mDao.updateMovie(movie, mId));
+				
+				return new ResponseEntity<ResponceStructure<Movie>>(str, HttpStatus.OK);
+			}
+			return null;
 		}
 		return null;
 	}
